Code Refactor, cleanup and bug fixes
