--- lib/constants.dart
@@ -1,10 +1,0 @@
-class Constants{
-  static const String Settings = 'Settings';
-  static const String Logout = 'Logout';
-
-  static const List<String> choices = <String>[
-    Logout
-  ];
-
-  static int PAGE_SIZE = 10;
-}
