+++ lib/screens/comment_input.dart
@@ -52,7 +52,7 @@ class CommentInput extends StatelessWidget {
           IconButton(
             icon: Icon(Icons.send),
             onPressed: () async {
-              if(_input.text.isEmpty) {
+              if (_input.text.isEmpty) {
                 return;
               }
               await _postComment(doctype, name, _input.text, authorEmail);
@@ -61,51 +61,14 @@ class CommentInput extends StatelessWidget {
           )
         ],
       ),
-      body: TextField(
-        autofocus: true,
-        controller: _input,
-        maxLines: 9999999,
+      body: Padding(
+        padding: const EdgeInsets.all(12.0),
+        child: TextField(
+          autofocus: true,
+          controller: _input,
+          maxLines: 9999999,
+        ),
       ),
     );
-    // return Container(
-    //   margin: EdgeInsets.only(bottom: 10),
-    //   decoration: BoxDecoration(
-    //       border: Border.all(color: Color.fromRGBO(209, 216, 221, 1))),
-    //   child: Column(
-    //     children: <Widget>[
-    //       Container(
-    //         height: 30,
-    //         decoration: BoxDecoration(
-    //             color: Color.fromRGBO(250, 251, 252, 1),
-    //             border: Border(
-    //                 bottom: BorderSide(
-    //                     width: 0.5, color: Color.fromRGBO(209, 216, 221, 1)))),
-    //         child: Row(
-    //           mainAxisAlignment: MainAxisAlignment.spaceBetween,
-    //           children: <Widget>[
-    //             Padding(
-    //               padding: EdgeInsets.only(left: 5),
-    //               child: Text(
-    //                 'Add a Comment',
-    //                 style: TextStyle(color: Color.fromRGBO(141, 153, 166, 1)),
-    //               ),
-    //             ),
-    //             FlatButton(
-    //               onPressed: () {
-    //                 FocusScope.of(context).unfocus();
-    //                 _postComment(doctype, name, input.text, authorEmail);
-    //               },
-    //               color: Color.fromRGBO(240, 244, 247, 1),
-    //               child: Text('Comment'),
-    //             )
-    //           ],
-    //         ),
-    //       ),
-    //       TextField(
-    //         controller: input,
-    //       ),
-    //     ],
-    //   ),
-    // );
   }
 }
