--- lib/widgets/file_picker.dart
@@ -1,124 +1,0 @@
-import 'dart:io';
-
-import 'package:flutter/material.dart';
-
-import 'package:flutter/services.dart';
-import 'package:file_picker/file_picker.dart';
-
-class FilePickerDemo extends StatefulWidget {
-  @override
-  _FilePickerDemoState createState() => new _FilePickerDemoState();
-}
-
-class _FilePickerDemoState extends State<FilePickerDemo> {
-  String _fileName;
-  String _path;
-  File file;
-  Map<String, String> _paths;
-  String _extension;
-  bool _loadingPath = false;
-  bool _multiPick = false;
-  FileType _pickingType;
-  TextEditingController _controller = new TextEditingController();
-
-  @override
-  void initState() {
-    super.initState();
-    _controller.addListener(() => _extension = _controller.text);
-  }
-
-  void _openFileExplorer() async {
-    setState(() => _loadingPath = true);
-    try {
-      
-      _path = null;
-      file = await FilePicker.getFile(
-          type: _pickingType,
-          allowedExtensions: (_extension?.isNotEmpty ?? false)
-              ? _extension?.replaceAll(' ', '')?.split(',')
-              : null);
-
-      print(file);
-
-      // var file = await FilePicker.getFile();
-      
-    } on PlatformException catch (e) {
-      print("Unsupported operation" + e.toString());
-    }
-    if (!mounted) return;
-    setState(() {
-      _loadingPath = false;
-      _fileName = _path != null
-          ? _path.split('/').last
-          : _paths != null ? _paths.keys.toString() : '...';
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return new MaterialApp(
-      home: new Scaffold(
-        appBar: new AppBar(
-          title: const Text('File Picker example app'),
-        ),
-        body: new Center(
-            child: new Padding(
-          padding: const EdgeInsets.only(left: 10.0, right: 10.0),
-          child: new SingleChildScrollView(
-            child: new Column(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: <Widget>[
-                new Padding(
-                  padding: const EdgeInsets.only(top: 50.0, bottom: 20.0),
-                  child: new RaisedButton(
-                    onPressed: () => _openFileExplorer(),
-                    child: new Text("Open file picker"),
-                  ),
-                ),
-                new Builder(
-                  builder: (BuildContext context) => _loadingPath
-                      ? Padding(
-                          padding: const EdgeInsets.only(bottom: 10.0),
-                          child: const CircularProgressIndicator())
-                      : _path != null || _paths != null
-                          ? new Container(
-                              padding: const EdgeInsets.only(bottom: 30.0),
-                              height: MediaQuery.of(context).size.height * 0.50,
-                              child: new Scrollbar(
-                                  child: new ListView.separated(
-                                itemCount: _paths != null && _paths.isNotEmpty
-                                    ? _paths.length
-                                    : 1,
-                                itemBuilder: (BuildContext context, int index) {
-                                  final bool isMultiPath =
-                                      _paths != null && _paths.isNotEmpty;
-                                  final String name = 'File $index: ' +
-                                      (isMultiPath
-                                          ? _paths.keys.toList()[index]
-                                          : _fileName ?? '...');
-                                  final path = isMultiPath
-                                      ? _paths.values.toList()[index].toString()
-                                      : _path;
-
-                                  return new ListTile(
-                                    title: new Text(
-                                      name,
-                                    ),
-                                    subtitle: Image.file(file),
-                                  );
-                                },
-                                separatorBuilder:
-                                    (BuildContext context, int index) =>
-                                        new Divider(),
-                              )),
-                            )
-                          : new Container(),
-                ),
-              ],
-            ),
-          ),
-        )),
-      ),
-    );
-  }
-}
